Move tagging UI code to binary
