+++ lib/entry/libimagentrytag/src/lib.rs
@@ -46,9 +46,7 @@ extern crate libimagstore;
 #[macro_use] extern crate libimagerror;
 
 pub mod error;
-pub mod exec;
 pub mod result;
 pub mod tag;
 pub mod tagable;
-pub mod ui;
